% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_lookup_nhs.R
\name{geog_lookup_nhs}
\alias{geog_lookup_nhs}
\title{A lookup table for NHS geographies}
\usage{
geog_lookup_nhs()
}
\value{
dataframe. A dataframe of lookup geographies, with codes and names.
}
\description{
A lookup table for NHS geographies
}
