% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdvizzer-package.R
\docType{package}
\name{imdvizzer-package}
\alias{imdvizzer}
\alias{imdvizzer-package}
\title{imdvizzer: Visualising UK Indices Of Multiple Deprivation}
\description{
Visualises the UK's Indices of Multiple Deprivation (IMD), in as many (or as few) ways as needed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThomUK/imdvizzer}
  \item Report bugs at \url{https://github.com/ThomUK/imdvizzer/issues}
}

}
\author{
\strong{Maintainer}: Tom Smith \email{tomsmith_uk@hotmail.com}

}
\keyword{internal}
