% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_nhs_imd.R
\name{lookup_nhs_imd}
\alias{lookup_nhs_imd}
\title{A lookup table of IMDs for NHS geographies}
\usage{
lookup_nhs_imd()
}
\value{
dataframe. A dataframe of IMD ranks and deciles, complete with NHS geographies, codes, and names.
}
\description{
A lookup table of IMDs for NHS geographies
}
