% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_imd_icb.R
\name{map_imd_icb}
\alias{map_imd_icb}
\title{Map IMD by ICB}
\usage{
map_imd_icb(icb_name, include_title = TRUE)
}
\arguments{
\item{icb_name}{character. The name of the ICB to map.}

\item{include_title}{logical. Decide whether to include the plot title. Default = TRUE}
}
\value{
list. A list containing a ggplot object, and the name of the ICB.
}
\description{
Map IMD by ICB
}
